Use cmake to build vlan python script
